Update StatusAction.php
I started using this bundle today and paid using stripe test card. One of the cards is supposed to return paid true and succeeded true but the bundle still say pending so i couldnt add credit to my user programaticaly. with this edit it works. Please double check before merge because im only a payum begginer and dont want to introduce a bug is a bundle as critical as this one for a business.